requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2
pytest==8.3.3
pytest-cov==5.0.0
bcrypt==4.0.1
